Split file manager into two
